# server/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DealForge API",
    "description": "AI-powered meeting intelligence and sales CRM API",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "/api", "description": "API server" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Read-only API key (generated in Settings → API Access). Available on Pro."
      }
    },
    "schemas": {
...
      "BufferData": {
        "type": "object",
        "properties": {
          "transcript": { "type": "object", "nullable": true },
          "notes": { "type": "object", "nullable": true },
          "meeting": { "type": "object", "nullable": true }
        }
      },
...
    }
  },
  "paths": {
...
    "/zoom/oauth/callback": {
      "get": {
        "summary": "Zoom OAuth callback",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "Authorization code from Zoom"
          }
        ],
        "responses": {
          "200": { "description": "Token exchange successful" },
          "400": { "description": "Missing authorization code" },
          "500": { "description": "Token exchange failed" }
        }
      }
    },
    "/zoom/webhook": {
      "post": {
        "summary": "Zoom webhook handler",
        "description": "Receives Zoom meeting events (started, ended) and URL validation",
        "responses": {
          "200": { "description": "Webhook processed" },
          "401": { "description": "Invalid signature" },
          "500": { "description": "Server configuration error" }
        }
      }
    },
    "/zoom/deauth": {
      "post": {
        "summary": "Zoom deauthorization webhook",
        "description": "Receives Zoom deauth events and cleans up user Zoom tokens",
        "responses": {
          "200": { "description": "Deauth processed" },
          "401": { "description": "Invalid signature" },
          "500": { "description": "Server configuration error" }
        }
      }
    },
    "/zoom/transcription": {
      "post": {
        "summary": "Store transcription segment",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["meetingId", "segment"],
                "properties": {
                  "meetingId": { "type": "string" },
                  "segment": { "type": "object" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Segment stored and broadcast" },
          "400": { "description": "Missing required fields" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/zoom/notes": {
      "post": {
        "summary": "Store quick notes from Zoom panel",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["meetingId", "note"],
                "properties": {
                  "meetingId": { "type": "string" },
                  "note": { "type": "object" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Note stored" },
          "400": { "description": "Missing required fields" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/zoom/buffer/{meetingId}": {
      "get": {
        "summary": "Pull buffered meeting data",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "meetingId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Buffered data",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/BufferData" } }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      },
      "delete": {
        "summary": "Clear buffered meeting data",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "meetingId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Buffer cleared" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
  }
}